+++ tarantool/src/error.rs
@@ -45,8 +45,15 @@ pub enum Error {
     #[error("failed to encode tuple: {0}")]
     Encode(#[from] Encode),
 
-    #[error("failed to decode tuple: {0}")]
-    Decode(#[from] rmp_serde::decode::Error),
+    #[error("Failed to decode tuple: {error} when decoding msgpack {} into rust type {expected_type}", crate::util::DisplayAsHexBytes(.actual_msgpack))]
+    Decode {
+        error: rmp_serde::decode::Error,
+        expected_type: String,
+        actual_msgpack: Vec<u8>,
+    },
+
+    #[error("Failed to decode tuple: {0}")]
+    DecodeRmpValue(#[from] rmp_serde::decode::Error),
 
     #[error("unicode string decode error: {0}")]
     Unicode(#[from] Utf8Error),
@@ -84,6 +91,17 @@ pub enum Error {
 
     #[error("space metadata not found")]
     MetaNotFound,
+}
+
+impl Error {
+    #[inline(always)]
+    pub fn decode<T>(error: rmp_serde::decode::Error, data: Vec<u8>) -> Self {
+        Error::Decode {
+            error,
+            expected_type: std::any::type_name::<T>().into(),
+            actual_msgpack: data,
+        }
+    }
 }
 
 impl From<rmp_serde::encode::Error> for Error {